requirements.mac.txt
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
docutils==0.14
idna==2.7
# Kivy==1.10.1
git+https://github.com/matham/kivy.git@async-loop
Kivy-Garden==0.1.4
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.1
pandas==0.23.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
tqdm==4.25.0
urllib3==1.23
